Refactored backend 'savedplants' routes
